+++ src/components/Header/CartQuantity/index.tsx
@@ -1,12 +1,12 @@
 import { CartActionErrorModal } from '@components/Fallbacks/index';
-import { useCartList } from '@hooks/index';
+import { useGetCartList } from '@hooks/index';
 
 import style from './style.module.css';
 
 const MAX_QUANTITY = 50;
 
 const CartQuantity = () => {
-  const { cartListMap, isSuccess, error } = useCartList();
+  const { cartListMap, isSuccess, error } = useGetCartList();
 
   const NONE_CART_LIST_LENGTH = 0;
   const cartListLength = cartListMap?.size ?? NONE_CART_LIST_LENGTH;
