+++ src/hooks/cart/index.ts
@@ -1,4 +1,4 @@
 export { default as useDeleteCartItem } from './useDeleteCartItem';
-export { default as useCartItemQuantity } from './useCartItemQuantity';
+export { default as useChangeCartItemQuantity } from './useChangeCartItemQuantity';
 export { default as useAddCartItem } from './useAddCartItem';
-export { default as useCartList } from './useCartList';
+export { default as useGetCartList } from './useGetCartList';
